fileaway: add presence management in terms of relay clients
If the availability file is empty (more precisely: contains
only whitespace) or missing, we retain the existing
behavior.

However, we add new functionality for the case
of a non-empty availability file.

Before going to the general case, let's discuss
a simple special case: if the availability file
consists of a never-matching regex (eg. '.^'),
then the plugin will completely rely on relay
connections: if a relay client is connected,
then it sets availabllity, if none is connected,
sets away.

The general behavior is as follows:
the content of the availability file is treated
as a regex.
- If a relay client is connected whose
  address does not match the regex, then
  we declare availability.
- If no relay client is connected, we
  declare away.
- If there are some relay clients and
  the addresses of all of them matches
  the regex, then we declare away if
  the availability file is expired, and
  availability if not yet.

Rationale: if Weehat is ran remotely,
it makes sense to equate relay client
presence with user's IRC presence.
So far, that is a completely different
presence management mechanism than "file away"
and could be implemented separately.

However, we want to improve on the above
idea and achieve that the IRC status
reflects the actual presence of the user
at the relay client's end, and for
that end we employ the availability file feature.

So if a client machine can arrange that
it updates the availability file on the Weechat
machine (eg. via SSH) precisely when user
presence is deceted (eg. by tracking monitor
status), then it can register itself as such
a smart client by writing its address into
the availability file.

Other clients, primarily smartphones, are less
capable and less sensitive: on a phone it might be
too cumbersome to set up such an updater script, but
on the other hand, there we have more reason to equate
relay connectedness with actual user presence, as a phone
can be regarded to be always present for the user (be
always at reach, able to notify any time).

The regex based mechanism can work out for both
kind of use cases.